+++ mocks/mock_avr_delay.c
@@ -5,9 +5,9 @@
 
 #include "mock_avr_delay.h"
 
-double total_us_delay = 0;
-double total_ms_delay = 0;
+float total_us_delay = 0;
+float total_ms_delay = 0;
 
-void _delay_us(double __us) { total_us_delay += __us; }
+void _delay_us(float __us) { total_us_delay += __us; }
 
-void _delay_ms(double __ms) { total_ms_delay += __ms; }
+void _delay_ms(float __ms) { total_ms_delay += __ms; }
