+++ mocks/mock_avr_delay.h
@@ -6,10 +6,10 @@
 #ifndef MOCK_AVR_DELAY_H
 #define MOCK_AVR_DELAY_H
 
-extern double total_us_delay;
-extern double total_ms_delay;
+extern float total_us_delay;
+extern float total_ms_delay;
 
-void _delay_us(double __us);
-void _delay_ms(double __ms);
+void _delay_us(float __us);
+void _delay_ms(float __ms);
 
 #endif  // MOCK_AVR_DELAY_H
